Post page is now ready to use and Entities and messages are updated - Ryan

+++ src/main/java/com/avosh/baseproject/repository/UserRepository.java
@@ -6,7 +6,7 @@ import com.avosh.baseproject.entity.SecUser;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends BaseRepository<SecUser, Integer> {
+public interface UserRepository extends BaseRepository<SecUser, Long> {
 
     SecUser findByEmail(String email);
